feat: Introduced sharp resizing function
Included a resize function which allows the
images to be resized. I also implemented a more
efficient way for the app to read images.
e.g. We first check if the thumb image exists,
if it exists we return it. If it does not exist we
read the full image, resize it, write it to disk
and then return it. I also added a way for the API
to determine if the width and height are numbers.
A HTTP 400 is returned if the width and height are
not numbers.