refactor: Implemented the POST endpoint to upload images
Added new server options when initiailizing.
Updated the index.html to handle the button press to upload and
to refresh after 200 is received from the server.
Added the createImage function in the imageController.ts to handle
the HTTP POST.
Installed CORS and use it in the server init function. This allows
for cross origin access requests, else the client cant talk to the
server.
Added a new sample data image data to be used when the server is
started and the createDirs option is set to true.
Added a bunch of tests as well.