refactor: Updated the /api/images/list endpoint
I added a new way to query for different directories, by adding
a URL param to the /api/images/list endpoint e.g.
/api/images/list/full.
This allows the client to fetch a list of assets from different
directories. I also added API tests for the updated endpoint.
I also included a fix for when I tested on Windows. The 0.0.0.0
host does not work on windows.