Feature: The server can now handle multiple file types and fit options
In this commit I changed the sample image to a jpg. In addition, the
server can now handle multiple file types. The images endpoint can
also accept an optional fit query parameter which allows the user
to specify the image fit properties of the image.
More information on this in the README.md.